# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RoVerse Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all platform users
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

    post:
      operationId: createUser
      tags: [users]
      summary: Add a user manually
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user role, subroles, or groups
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    delete:
      operationId: deleteUser
      tags: [users]
      summary: Remove a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /planning/teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List all teams with their members
      responses:
        "200":
          description: Teams list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"

    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a team (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInput"
      responses:
        "201":
          description: Created team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /planning/teams/{id}:
    patch:
      operationId: updateTeam
      tags: [teams]
      summary: Update a team (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamUpdate"
      responses:
        "200":
          description: Updated team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"

    delete:
      operationId: deleteTeam
      tags: [teams]
      summary: Delete a team (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /planning/teams/{id}/members:
    post:
      operationId: addTeamMember
      tags: [teams]
      summary: Add a member to a team (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddTeamMemberInput"
      responses:
        "201":
          description: Added member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /planning/teams/{id}/members/{userId}:
    delete:
      operationId: removeTeamMember
      tags: [teams]
      summary: Remove a member from a team (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
...
components:
  schemas:
...
    User:
      type: object
      required: [id, robloxId, robloxUsername, role, createdAt]
      properties:
        id:
          type: integer
        robloxId:
          type: string
        robloxUsername:
          type: string
        robloxDisplayName:
          type: ["string", "null"]
        robloxAvatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, collaborator]
        subroles:
          type: array
          items:
            type: string
        groups:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    UserInput:
      type: object
      required: [robloxId, role]
      properties:
        robloxId:
          type: string
        role:
          type: string
          enum: [admin, collaborator]
        subroles:
          type: array
          items:
            type: string
        groups:
          type: array
          items:
            type: string
    UserUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [admin, collaborator]
        subroles:
          type: array
          items:
            type: string
        groups:
          type: array
          items:
            type: string
...
    Team:
      type: object
      required: [id, name, color, createdAt, members]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: string
        createdAt:
          type: string
          format: date-time
        members:
          type: array
          items:
            $ref: "#/components/schemas/TeamMember"
    TeamInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        color:
          type: string
    TeamUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: string
    TeamMember:
      type: object
      required: [userId, username, role]
      properties:
        userId:
          type: integer
        username:
          type: string
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, collaborator]
    AddTeamMemberInput:
      type: object
      required: [userId]
      properties:
        userId:
          type: integer
...