# lib/api-zod/src/generated/types/team.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RoVerse Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { TeamMember } from './teamMember';

export interface Team {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  color: string;
  createdAt: Date;
  members: TeamMember[];
}