# lib/api-zod/src/generated/types/teamMember.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RoVerse Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { TeamMemberRole } from './teamMemberRole';

export interface TeamMember {
  userId: number;
  username: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  role: TeamMemberRole;
}